Suggestion for roadmap doc
My idea here, is that this is a frozen point in time version of the current README, so it will never be updated. It is *only* for the purposes of giving to the NIH in conjunction with the level 0 popout, in order to show our thoughts and roadmap towards other levels. 

Important edits:

- changed 'final' to current
- lots of edits in lines 57-74 for added explanation
- tried to remove the idea that we have multiple models